backend/requirements.txt
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
xgboost==2.0.3
joblib==1.3.2
torch==2.2.1
transformers==4.38.2
accelerate==0.27.2
requests==2.31.0
wbgapi==1.0.12
httpx==0.27.0
scipy==1.12.0
fastapi==0.110.0
uvicorn==0.28.0
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.3
tqdm==4.66.2
python-dateutil==2.9.0
